requirements-dev.txt
Django==1.4.3
MySQL-python==1.2.4
South==0.7.6
amqp==1.0.8
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.21
celery==3.0.15
django-celery==3.0.11
git+https://github.com/marbindrakon/django-eveigb
django-infinite-memcached==0.1
eveapi==1.2.6
feedparser==5.1.3
kombu==2.5.6
python-dateutil==1.5
python-memcached==1.48
pytz==2012j
sleekxmpp==1.1.11
wsgiref==0.1.2
django-nose==1.1
Sphinx==1.2b1
djangorestframework==3.1.1
djangorestframework-jwt==1.6.0
